HealthyFins/Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow-cpu==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0